RFC 68: Make C++11 a build requirement (https://github.com/OSGeo/gdal/pull/270)
- configure: make C++11 a requirement. Add a --with-cpp14 option
- cpl_port.h: make C++11 / VS>=2013 a requirement
- a few exceptions in autotest for weird failures with mingw64/wine
- Travis-CI: remove mingw target using ancient non-C++11 mingw32 compiler
- Travis-CI: move precise_32bit to trusty_32bit
- Travis-CI: update Android toolchain
- Travis-CI: update mingw_w64 to trusty
- appveyor.yml: use current MSVC version to build Python bindings
- Python bindings: automatically add -std=c++11 for Unix builds



git-svn-id: https://svn.osgeo.org/gdal/trunk@40984 f0d54148-0727-0410-94bb-9a71ac55c965

+++ autotest/pyscripts/test_gdal2tiles.py
@@ -76,6 +76,10 @@ def test_gdal2tiles_py_zoom_option():
     if script_path is None:
         return 'skip'
 
+    # Issue with multiprocessing in the chroot
+    if os.environ.get('BUILD_NAME', '') == 'trusty_32bit':
+        return 'skip'
+
     shutil.rmtree('tmp/out_gdal2tiles_smallworld', ignore_errors=True)
 
     # Because of multiprocessing, run as external process, to avoid issues with
